added wine review rating to filter with star emoticon; working on editing user account from profile
