+++ public/index.css
@@ -107,6 +107,11 @@ button {
 .search-bars-container {
   text-align: center;
 }
+
+.material-symbols-outlined {
+  font-variation-settings: "FILL" 100, "wght" 400, "GRAD" 0, "opsz" 48;
+}
+
 /* wine list */
 
 /* winefeed */
